Authentication using service account.
